+++ Subject/src/main/java/com/example/Subject/entity/Student.java
@@ -1,6 +1,6 @@
 package com.example.Subject.entity;
 
-import jakarta.persistence.*;
+
 import lombok.*;
 
 import java.util.List;
@@ -10,7 +10,7 @@ import java.util.List;
 @NoArgsConstructor
 public class Student {
 
-    private Long id;
+    private int id;
     private String name;
     private String phoneNumber;
     private String address;
